requirements.txt
annotated-types==0.7.0
anyio==4.8.0
bcrypt==4.0.1
cffi==1.17.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.0
distlib==0.3.8
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
fastapi==0.115.8
filelock==3.13.4
greenlet==3.1.1
h11==0.14.0
httptools==0.6.1
idna==3.10
MarkupSafe==3.0.2
passlib==1.7.4
platformdirs==4.2.0
pyasn1==0.4.8
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PyMySQL==1.1.1
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.15.0
python-dotenv==1.0.1
python-jose==3.4.0
python-multipart==0.0.20
PyYAML==6.0.1
rsa==4.9
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.37
starlette==0.45.3
typing_extensions==4.12.2
uvicorn==0.34.0
virtualenv==20.25.1
watchfiles==0.21.0
websockets==12.0
Werkzeug==3.1.3
